requirements-common.txt
namex>=0.0.8
black>=22
flake8
isort
pytest
numpy
scipy
pandas
absl-py
requests
h5py
ml-dtypes
protobuf
google
tensorboard-plugin-profile
rich
build
optree
pytest-cov
